feat: Add createFaviconFromH1 function
This commit adds a new function called createFaviconFromH1. This function creates a favicon from the text content of the h1 element on the page. It sets the text color to #28b42f and uses a 16x16 canvas to generate the favicon image. The resulting image is then set as the href for the link element with rel="icon" in the document's head, creating or updating the favicon.